Place items radially
treeItems are now placed radially, in concentric circles according to
how many steps away from the final product each item is. Each item gets
a certain arc of the circle depending on how many children it has, then
each of its children is spread out within that arc proportionally to how
many children it has, etc.

This display method was designed for the Soul of Eternity from Fargo's
Mod, an item with an intricate crafting tree, not the Zenith with 10 top
level ingredients. For that reason, the Zenith's crafting tree appears
crowded, and the anti-collision code is used to keep items at least
their socialDistance apart from each other.