Zoom and pan with mouse
Zooming and panning around have been implemented using the OpenGL
camera. Because the mouse position is now no longer rigidly related to
positions on the canvas, some maths is used to determine where the mouse
is actually pointing. This is used to check for mouse collision for
highlighting items' socialDistance circles, and from now on mouseX and
mouseY will only be used for UI elements.