Select crafting tree to display
When loading, there is now a menu to choose which item to load the
crafting tree of. Once loaded, if the user presses escape they are able
to choose a different item.

The definitions of these items, as well as craftingStations, have been
moved to crafting-tree-items.json, and in-game-items.json has had its
element renamed to "terraria" to prepare for selective sprite loading.