Only load required sprites
To avoid unnecessarily long load times, only the sprites that are
required for the selectableItems and craftingStations are loaded
initially, then when the items are being loaded recursively, only the
sprites that will be used on the specific items being loaded. This is
achieved by simply not giving the inGameItems a sprite at first, looking
through the loaded ones to see which don't have a sprite yet, and
loading those. This is performant, and offers more granularity (and thus
shorter load times) than the itemSets approach. The "Loading sprites"
screen now also shows the progress being made, so users don't think
they're waiting for nothing.