Add Shadows of Abaddon and Thorium items
All of the Shadows of Abaddon and Thorium selectableItems have been
added, meaning all of the items have been implemented except for the two
Souls of Eternity. The current systems for arranging and displaying
crafting trees work fine for all of the existing items, albeit with poor
performance when viewing some more complex trees (using a 3070 on
Firefox), though some rewriting and optimisation will likely be required
for the Soul of Eternity to be viewable at a decent framerate on average
hardware.