Add Soul of the Universe temporarily
The Soul of the Universe, with and without mod add-ons, has been added
as a temporary stand in for the Soul of Eternity. Its recipe is too
crowded for the radial crafting tree to work, meaning a new display
method will be required to display the Soul of Eternity in a single
diagram without it looking like a big mess.

To improve performance somewhat, images and arrows are only drawn if
they are visible on screen, which will certainly be useful with the
hundreds of items that make up the Soul of Eternity crafting tree.