Display crafting tree in different layouts
The crafting tree can now be displayed in five different layouts; a
standard crafting tree layout with the final product at each of the four
cardinal directions, and the existing radial tree layout. A few of the
more crowded crafting trees have had the radial layout disabled, such as
the Drataliornus and the Soul of the Universe, meaning that itemSpacing
no longer needs to start at ~1000 for them. This solution is a little
inelegant, and the larger crafting trees are more or less useless, but
it works and doesn't look terrible.

The item display functions have also been rewritten, so that items are
updated, then drawn one element at a time, every frame.

At this point, once the Soul of Eternity is implemented, and mouseover
detection is rewritten using a quadtree (if needed), this project should
be complete.