Add Soul of Eternity and all ingredients
The Soul of Eternity from Fargo's Soul Mod has been added, as has its
mod add-ons version, including items from Calamity, Thorium and Shadows
of Abaddon. Over 1200 new items have been added, including sprites and
entries in in-game-items.json, more than doubling the total amount.
Loading the Soul of Eternity tree takes a while, and notably spends many
seconds seemingly loading the first few sprites. This only happens the
first time, so it may be loadItemRecursive running. A new loading screen
(or an additional line on the loading sprites screen) may be needed to
explain that the items are being loaded.

The crafting tree generator is functionally complete. As tModLoader for
1.4 is already in alpha, an update to all of the current recipes might
be required soon, switching all the 1.3 recipes for their 1.4
equivalents, both of which were already added to in-game-items for this
very reason.